Add minimal support for RPM macros in Name/Version/Release fields
Only "%name" and "%{name}" substitutions are handled, and only for
macros which are defined in the same spec file (the only exception
currently is %nil, which is predefined to the empty string).  "%%" is
expanded to "%"; if any other uses of the '%' character are found, the
field value is left as is (which will cause gear(1) to fail if such
value is used in keyword substitutions).

Conditionals in the spec file are not interpreted - all "%define" and
"%global" operators found in the spec file are considered, and the
last definition is used in case of multiple definitions, which may not
match actual RPM behavior.

Signed-off-by: Sergey Vlasov <vsu@altlinux.ru>