gear-sh-functions (get_filename_from_rules): drop second argument
The second argument of get_filename_from_rules() was an alternative name
of the directive name (passed as the first argument) for use in error
diagnostic messages.  However, the directive name itself is enough, so
the second argument is no more.

This change also updates all get_filename_from_rules() users.